read file from end and color in console

+++ src/test_ut_string.c
@@ -1,6 +1,7 @@
 #include "test_ut_string.h"
 #include "utstring.h"
-
+#include <string.h>
+#include <stdio.h>
 
 void run_test_UTString(void)
 {
@@ -8,7 +9,7 @@ void run_test_UTString(void)
 	char f[10];
 	utstring_new(s);
 	utstring_printf(s, "hello world!");
-	printf("char : %s  length :%d \n", utstring_body(s),strlen(utstring_body(s)));
+	printf("char : %s  length :%d \n", utstring_body(s),(int)strlen(utstring_body(s)));
 	 f[9]='\0' ;
 	strcpy(f, "salam");
 	utstring_clear(s);
